Made changes as requested by the reviwer
